assets+network: validate untrusted input, fix overflow/UB, peer-key collision
Assets (untrusted file parsing):
- GLB loader used JSON/BIN chunk lengths directly as allocation sizes with no
  bound check; now validated against remaining file size.
- TGA loader reserved the decode buffer (up to ~17 GB) before any size check;
  dimensions are now capped to a sane maximum.
- AssetPackage TOC read entry_count and per-entry path_len unbounded; both are
  now validated against the file size before allocating.
- glTF importer computed accessor byte offsets in u32, which could wrap and
  slip past the bounds check into an out-of-bounds read; promoted to size_t.

Network:
- BitReader::read_string/read_bytes reserved an attacker-controlled length up
  front (memory-exhaustion DoS); reserve is capped to bytes actually remaining.
- UdpTransport::make_peer_key shifted the IP by only 16 bits, discarding the
  top 16 bits so distinct addresses aliased; now shifts 32 and checks inet_pton.
- Snapshot interpolation reinterpret_cast'd unaligned u8 buffers to float*
  (UB); now copies via memcpy.

perf: Profiler::last_frame returned a pointer into the history buffer after
releasing the lock (dangling on concurrent end_frame); returns by value now.

+++ engine/assets/src/gltf_importer.cpp
@@ -283,10 +283,13 @@ std::vector<float> read_accessor_floats(const GltfScene& scene, u32 accessor_ind
     std::vector<float> result(static_cast<size_t>(acc.count) * comp_count);
 
     for (u32 i = 0; i < acc.count; ++i) {
-        u32 offset = bv.byte_offset + acc.byte_offset + i * stride;
+        // Compute offsets in size_t: u32 arithmetic on attacker-controlled
+        // byte_offset/stride can wrap and slip past the bounds check below.
+        size_t offset = static_cast<size_t>(bv.byte_offset) + acc.byte_offset
+                      + static_cast<size_t>(i) * stride;
         for (u32 c = 0; c < comp_count; ++c) {
-            u32 byte_pos = offset + c * comp_sz;
-            if (byte_pos + comp_sz > static_cast<u32>(buf.size())) {
+            size_t byte_pos = offset + static_cast<size_t>(c) * comp_sz;
+            if (byte_pos + comp_sz > buf.size()) {
                 result[static_cast<size_t>(i) * comp_count + c] = 0.0f;
                 continue;
             }
@@ -320,8 +323,9 @@ std::vector<u32> read_accessor_indices(const GltfScene& scene, u32 accessor_inde
 
     std::vector<u32> result(acc.count);
     for (u32 i = 0; i < acc.count; ++i) {
-        u32 offset = bv.byte_offset + acc.byte_offset + i * stride;
-        if (offset + comp_sz > static_cast<u32>(buf.size())) { result[i] = 0; continue; }
+        size_t offset = static_cast<size_t>(bv.byte_offset) + acc.byte_offset
+                      + static_cast<size_t>(i) * stride;
+        if (offset + comp_sz > buf.size()) { result[i] = 0; continue; }
         const u8* ptr = buf.data() + offset;
         switch (acc.component_type) {
             case 5121: result[i] = static_cast<u32>(*ptr); break;
